nav update
Updated the Nav to have a taller height of 85px so that it is the same height as the grid that lives in the Nav. I further updated the flex of the items so that it fits better and, when view model if changed, the Nav reacts in a cleaner way.